fix: proper datetime display

+++ app/src/main/java/com/example/weather/dto/SysResponse.kt
@@ -3,6 +3,6 @@ package com.example.weather.dto
 import com.google.gson.annotations.SerializedName
 
 data class SysResponse (
-    @SerializedName("sunrise") val sunrise: Int,
-    @SerializedName("sunset") val sunset: Int
+    @SerializedName("sunrise") val sunrise: Long,
+    @SerializedName("sunset") val sunset: Long
 )
